Add HTML as an output format (#247)
Generate a table of registers/fields
Fix BMP280 codegen based on signed values/data sizes

See #244

+++ test/sampleData/arduino/BMP280.h
@@ -168,7 +168,7 @@ class BMP280 {
          * Reads the atmospheric pressure
 
          */
-        short pressureasRaw();
+        long pressureasRaw();
 
         /**
          * Reads the temperature
@@ -179,7 +179,7 @@ class BMP280 {
          * Reads the temperature
 
          */
-        short temperatureasRaw();
+        long temperatureasRaw();
 
 
     private:
