+++ test/sampleData/esp32/BMP280.h
@@ -167,7 +167,7 @@ class BMP280 {
          * Reads the atmospheric pressure
 
          */
-        short pressureasRaw();
+        long pressureasRaw();
 
         /**
          * Reads the temperature
@@ -178,7 +178,7 @@ class BMP280 {
          * Reads the temperature
 
          */
-        short temperatureasRaw();
+        long temperatureasRaw();
 
 
     private:
